refactor(types): promote execution-payload schema into common/types
  Phase 1a of the M6 plan (docs/plans/engine-api-integration.md, also
  updated in this commit). The canonical block-component types —
  ExecutionPayloadV3, Withdrawal, HexBytes, and the hex_* serde helpers —
  move from the engine-client crate into the foundational types crate,
  where the Lean BlockBody can later embed them directly.

  ethlambda-ethrex-client's public API stays stable through re-exports.
  No SSZ derives yet; those land in Phase 2 alongside the BlockBody embed.
